split PythonCppInterop.h to decouple cpp declarations from pybind11 types
