+++ python/runtime/interop/CMakeLists.txt
@@ -13,7 +13,7 @@ target_include_directories(cudaq-python-interop PRIVATE
     ${pybind11_INCLUDE_DIRS}
 )
 target_link_libraries(cudaq-python-interop PRIVATE pybind11::module cudaq)
-install (FILES PythonCppInterop.h DESTINATION include/cudaq/python/)
+install (FILES PythonCppInterop.h PythonCppInteropDecls.h DESTINATION include/cudaq/python/)
 
 install(TARGETS cudaq-python-interop EXPORT cudaq-python-interop-targets DESTINATION lib)
